requirements.txt
praw==7.7.1
python-dotenv==1.0.0
agno==0.1.0
requests==2.31.0
google-generativeai==0.8.3
schedule==1.2.0
prawcore==2.4.0
typing-extensions>=4.0.0
pydantic>=2.0.0
